+++ iocore/net/quic/QUICTLS_openssl.cc
@@ -222,7 +222,7 @@ QUICTLS::_decrypt(uint8_t *plain, size_t &plain_len, size_t max_plain_len, const
 }
 
 bool
-QUICTLS::_encrypt_pn(uint8_t *protected_pn, size_t &protected_pn_len, const uint8_t *unprotected_pn, size_t unprotected_pn_len,
+QUICTLS::_encrypt_pn(uint8_t *protected_pn, uint8_t &protected_pn_len, const uint8_t *unprotected_pn, uint8_t unprotected_pn_len,
                      const uint8_t *sample, const KeyMaterial &km, const EVP_CIPHER *aead) const
 {
   EVP_CIPHER_CTX *ctx = EVP_CIPHER_CTX_new();
@@ -245,7 +245,7 @@ QUICTLS::_encrypt_pn(uint8_t *protected_pn, size_t &protected_pn_len, const uint
 }
 
 bool
-QUICTLS::_decrypt_pn(uint8_t *unprotected_pn, size_t &unprotected_pn_len, const uint8_t *protected_pn, size_t protected_pn_len,
+QUICTLS::_decrypt_pn(uint8_t *unprotected_pn, uint8_t &unprotected_pn_len, const uint8_t *protected_pn, uint8_t protected_pn_len,
                      const uint8_t *sample, const KeyMaterial &km, const EVP_CIPHER *aead) const
 {
   EVP_CIPHER_CTX *ctx = EVP_CIPHER_CTX_new();
